requirements.txt
streamlit
pandas
pandasai
python-dotenv
matplotlib
seaborn

-e .